requirements.txt
asgiref==3.2.3
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.3
blinker==1.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
Django==3.0.3
django-admin==2.0.0
django-excel-response2==2.0.8
django-six==1.0.4
flake8==3.8.3
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.7.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
klaviyo==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
pytz==2019.3
requests==2.24.0
screen==1.0.1
sentry-sdk==0.16.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.3.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.11.2
xlwt==1.3.0
zipp==3.1.0
